+++ package.json
@@ -14,6 +14,7 @@
   },
   "dependencies": {
     "body-parser": "^1.18.3",
+    "csv-write-stream": "^2.0.0",
     "express": "^4.16.4",
     "faker": "^4.1.0",
     "jquery": "^3.3.1",
